Move message defs back into kernel.ts
